backend/requirements.txt
flask
flask-cors
PyJWT[crypto]
python-dotenv
requests
